+++ CHANGELOG.md
@@ -39,6 +39,7 @@ The following changes have been completed, and will be included in the next tagg
     - added `Filepath.parse()`
     - added `Filepath.relative()`
     - added `Filepath.resolve()`
+    - added `Filepath.toNamespacedPath()`
   - added `MakeFilepathOptions`
   - added `isFilepathData()`
   - added `makeFilepath()`